Move analytics lib to stepper
