+++ client/landing/stepper/utils/analytics/types.ts
